requirements.txt
Flask>=2.0.0
PyYAML>=5.4.0
python-dotenv>=0.19.0
urllib3>=1.26.0
